playing with tests

+++ runtests.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
 function quitting(){
-  kill -9 $(<test.pid)
-  rm -f test.pid
+  kill -9 $(<test.pid) >/dev/null 2>&1
+#  rm -f test.pid
   exit 1
 }
 
@@ -30,16 +30,21 @@ if [ "q$2" == "q" ]; then
 # else
 # 	echo "Checking for test-methods matching $m"
 fi
-echo "Cleaning up..."
-
-mvn clean > /dev/null
+if [ "$nodel" -eq 0 ]; then
+  echo "Cleaning up..."
+  mvn clean > /dev/null
+fi
 
 rg -l "@Test" | grep ".java" >files.lst
 rg -l "@ParameterizedTest" | grep ".java" >>files.lst
 
 sort -u files.lst | grep "$p" | sed -e 's!/!.!g' | sed -e 's/src.test.java//g' | sed -e 's/.java$//' | sed -e 's/^\.//'>files.txt
 cnt=$(wc -l < files.txt|tr -d ' ')
-testMethods=$(egrep "@Test|@ParameterizedTest" $(<files.lst)| cut -f2 -d: | grep -v '^ *//' | wc -l)
+if [ "$cnt" -eq 0 ]; then
+  echo "no matching class found for $p"
+  exit 1
+fi
+testMethods=$(egrep "@Test|@ParameterizedTest" $(<files.lst)| cut -f2 -d: | grep -vc '^ *//')
 if [ "$nodel" -eq 0 ]; then
   rm -rf test.log
   mkdir test.log
@@ -53,10 +58,10 @@ for t in $(<files.txt); do
   tm=$(date +%s)
   if [ "$m" == "." ]; then
     mvn -Dsurefire.useFile=false test -Dtest="$t" >test.log/"$t".log &
-    echo "$&" > test.pid
+    echo $! > test.pid
   else
     mvn -Dsurefire.useFile=false test -Dtest="$t#$m" >"test.log/$t.log" &
-    echo "$&" > test.pid
+    echo $! > test.pid
   fi
   while true; do
     clear
@@ -87,7 +92,10 @@ for t in $(<files.txt); do
     tail -n 10 test.log/"$t".log
     echo "----------"
     echo "Failed tests"
-    egrep "] Running |Tests run: " test.log/* | grep -B1 FAILURE | cut -f2 -d']' |grep -v "Tests run: " | sed -e 's/Running //' | grep -v -- '--' | pr -l1 -3 -t -w 280 || echo "none"
+    ./getFailedTests.sh | pr -t -3 -w280
+    # egrep "] Running |Tests run: " test.log/* | grep -B1 FAILURE | cut -f2 -d']' |grep -v "Tests run: " | sed -e 's/Running //' | grep -v -- '--' | pr -l1 -3 -t -w 280 || echo "none"
+
+
     # egrep "] Running |Tests run: " test.log/* | grep -B1 FAILURE | cut -f2 -d']' |grep -v "Tests run: " | sed -e 's/Running //' | grep -v -- '--'  || echo "none"
     # egrep "] Running |Tests run:" test.log/* | grep -B1 FAILURE | cut -f2 -d']' || echo "none"
     jobs >/dev/null
